BE/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary